fix(storybook): configure .storybook/preview.tsx for fix conflict with MUI
